update on the codes
Added 30 more lines at the beginning of codes for creating the SQLite database, its tables according to the information given in the IMDB website and inserting the rows. Upon downloading the tsv files from IMDB and changing the path in the codes, it should be able to run smoothly and give you the results and plots.